created navigation dir with navbar and navlink inside
