+++ frontend/src/components/navigation/NavLink.module.css
@@ -1,0 +1,15 @@
+a.current {
+    color: var(--color-dark-green);
+}
+
+a {
+    color: var(--color-white);
+    font-family: Inter;
+    font-size: 20px;
+    font-weight: 600;
+    line-height: 20px;
+    letter-spacing: 0px;
+    text-align: left;
+    padding-bottom: 5px;
+    text-decoration: none;
+}
