+++ frontend/src/components/navigation/Navbar.module.css
@@ -30,19 +30,3 @@ img {
     right: 3rem;
     top: 22px;
 }
-
-a.current {
-    color: var(--color-dark-green);
-}
-
-a {
-    color: var(--color-white);
-    font-family: Inter;
-    font-size: 20px;
-    font-weight: 600;
-    line-height: 20px;
-    letter-spacing: 0px;
-    text-align: left;
-    padding-bottom: 5px;
-    text-decoration: none;
-}
