+++ frontend/src/pages/Profile.tsx
@@ -1,4 +1,4 @@
-import Navbar from '../components/Navbar'
+import Navbar from '../components/navigation/Navbar'
 
 const Profile = () => {
     return (
